Add files via upload
playing with the color palette and rotating the text. I added more dialogue. I realized the way I made italics the first time didn’t quite work out because it resulted in this format (//italics//) rather than (italics). At the time I thought, I guess this is just how it will look in this program, now I realized it was because it was coded wrong.
I also caught typos. It’s easy to make overlook typos when doing this.
Formatted color on last page.